Use vi mode for shell and remove bash prompt config
In vi mode it's nice to show the current mode in the prompt. The easiest way is to use the existing prompt and configure it via inputrc. Anyway it's no longer possible to show the current result of the last command in the prompt (especially based by colour of my prompt) so I removed that code.